rename package and file to consensus_channel
