+++ channel/consensus_channel/consensus_channel.go
@@ -1,4 +1,4 @@
-package ledger
+package consensus_channel
 
 import (
 	"fmt"
